requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
gunicorn==20.1.0
flask-cors==3.0.10
Pillow==10.0.0
lxml==4.9.3
Flask-SocketIO==5.3.4
pymongo[srv]==4.8.0
python-dotenv==0.21.1
EbookLib==0.18
faiss-cpu==1.8.0
transformers==4.41.2
psutil==6.0.0
openai==1.3.7
beautifulsoup4==4.12.3
tiktoken==0.5.2
cryptography==41.0.7
numpy==1.26.4
unstructured[openai]==0.14.9